Add e2e tests for no-proxy with UMA
