+++ e2e/e2e_test.go
@@ -236,331 +236,3 @@ var _ = Describe("Code Flow PKCE login/logout", func() {
 		Expect(resp.StatusCode()).To(Equal(http.StatusSeeOther))
 	})
 })
-
-var _ = Describe("UMA Code Flow authorization", func() {
-	var portNum string
-	var proxyAddress string
-	var umaCookieName = "TESTUMACOOKIE"
-
-	BeforeEach(func() {
-		server := httptest.NewServer(&testsuite.FakeUpstreamService{})
-		portNum = generateRandomPort()
-		proxyAddress = "http://localhost:" + portNum
-		osArgs := []string{os.Args[0]}
-		proxyArgs := []string{
-			"--discovery-url=" + idpRealmURI,
-			"--openid-provider-timeout=120s",
-			"--listen=" + "0.0.0.0:" + portNum,
-			"--client-id=" + umaTestClient,
-			"--client-secret=" + umaTestClientSecret,
-			"--upstream-url=" + server.URL,
-			"--no-redirects=false",
-			"--enable-uma=true",
-			"--cookie-uma-name=" + umaCookieName,
-			"--skip-access-token-clientid-check=true",
-			"--skip-access-token-issuer-check=true",
-			"--openid-provider-retry-count=30",
-			"--secure-cookie=false",
-		}
-
-		osArgs = append(osArgs, proxyArgs...)
-		startAndWait(portNum, osArgs)
-	})
-
-	When("Accessing resource, where user is allowed to access", func() {
-		It("should login with user/password and logout successfully", func(ctx context.Context) {
-			var err error
-			rClient := resty.New()
-			resp := codeFlowLogin(rClient, proxyAddress+umaAllowedPath, http.StatusOK)
-			Expect(resp.Header().Get("Proxy-Accepted")).To(Equal("true"))
-
-			body := resp.Body()
-			Expect(strings.Contains(string(body), umaCookieName)).To(BeTrue())
-
-			By("Accessing not allowed path")
-			resp, err = rClient.R().Get(proxyAddress + umaForbiddenPath)
-			body = resp.Body()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusForbidden))
-			Expect(strings.Contains(string(body), umaCookieName)).To(BeFalse())
-
-			resp, err = rClient.R().Get(proxyAddress + "/oauth/logout")
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusOK))
-
-			rClient.SetRedirectPolicy(resty.NoRedirectPolicy())
-			resp, _ = rClient.R().Get(proxyAddress + umaAllowedPath)
-			Expect(resp.StatusCode()).To(Equal(http.StatusSeeOther))
-		})
-	})
-
-	When("Accessing resource, which does not exist", func() {
-		It("should be forbidden without permission ticket", func(ctx context.Context) {
-			rClient := resty.New()
-			resp := codeFlowLogin(rClient, proxyAddress+umaNonExistentPath, http.StatusForbidden)
-
-			body := resp.Body()
-			Expect(strings.Contains(string(body), umaCookieName)).To(BeFalse())
-		})
-	})
-
-	When("Accessing resource, which exists but user is not allowed and then allowed resource", func() {
-		It("should be forbidden and then allowed", func(ctx context.Context) {
-			var err error
-			rClient := resty.New()
-			resp := codeFlowLogin(rClient, proxyAddress+umaForbiddenPath, http.StatusForbidden)
-
-			body := resp.Body()
-			Expect(strings.Contains(string(body), umaCookieName)).To(BeFalse())
-
-			By("Accessing allowed resource")
-			resp, err = rClient.R().Get(proxyAddress + umaAllowedPath)
-			body = resp.Body()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusOK))
-			Expect(strings.Contains(string(body), umaCookieName)).To(BeFalse())
-
-			By("Accessing allowed resource one more time, checking uma cookie set")
-			resp, err = rClient.R().Get(proxyAddress + umaAllowedPath)
-			body = resp.Body()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusOK))
-			Expect(strings.Contains(string(body), umaCookieName)).To(BeTrue())
-		})
-	})
-})
-
-var _ = Describe("UMA Code Flow authorization with method scope", func() {
-	var portNum string
-	var proxyAddress string
-	var umaCookieName = "TESTUMACOOKIE"
-
-	BeforeEach(func() {
-		server := httptest.NewServer(&testsuite.FakeUpstreamService{})
-		portNum = generateRandomPort()
-		proxyAddress = "http://localhost:" + portNum
-		osArgs := []string{os.Args[0]}
-		proxyArgs := []string{
-			"--discovery-url=" + idpRealmURI,
-			"--openid-provider-timeout=120s",
-			"--listen=" + "0.0.0.0:" + portNum,
-			"--client-id=" + umaTestClient,
-			"--client-secret=" + umaTestClientSecret,
-			"--upstream-url=" + server.URL,
-			"--no-redirects=false",
-			"--enable-uma=true",
-			"--enable-uma-method-scope=true",
-			"--cookie-uma-name=" + umaCookieName,
-			"--skip-access-token-clientid-check=true",
-			"--skip-access-token-issuer-check=true",
-			"--openid-provider-retry-count=30",
-			"--secure-cookie=false",
-			"--verbose=true",
-			"--enable-logging=true",
-		}
-
-		osArgs = append(osArgs, proxyArgs...)
-		startAndWait(portNum, osArgs)
-	})
-
-	When("Accessing resource, where user is allowed to access and then not allowed resource", func() {
-		It("should login with user/password, don't access forbidden resource and logout successfully", func(ctx context.Context) {
-			var err error
-			rClient := resty.New()
-			resp := codeFlowLogin(rClient, proxyAddress+umaMethodAllowedPath, http.StatusOK)
-			Expect(resp.Header().Get("Proxy-Accepted")).To(Equal("true"))
-
-			body := resp.Body()
-			Expect(strings.Contains(string(body), umaCookieName)).To(BeTrue())
-
-			By("Accessing not allowed method")
-			resp, err = rClient.R().Post(proxyAddress + umaMethodAllowedPath)
-			body = resp.Body()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusForbidden))
-			Expect(strings.Contains(string(body), umaCookieName)).To(BeFalse())
-
-			resp, err = rClient.R().Get(proxyAddress + "/oauth/logout")
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusOK))
-
-			rClient.SetRedirectPolicy(resty.NoRedirectPolicy())
-			resp, _ = rClient.R().Get(proxyAddress + umaAllowedPath)
-			Expect(resp.StatusCode()).To(Equal(http.StatusSeeOther))
-		})
-	})
-})
-
-var _ = Describe("UMA no-redirects authorization with forwarding client credentials grant", func() {
-	var portNum string
-	var proxyAddress string
-	var fwdPortNum string
-	var fwdProxyAddress string
-
-	BeforeEach(func() {
-		server := httptest.NewServer(&testsuite.FakeUpstreamService{})
-		portNum = generateRandomPort()
-		fwdPortNum = generateRandomPort()
-		proxyAddress = "http://localhost:" + portNum
-		fwdProxyAddress = "http://localhost:" + fwdPortNum
-		osArgs := []string{os.Args[0]}
-		fwdOsArgs := []string{os.Args[0]}
-		proxyArgs := []string{
-			"--discovery-url=" + idpRealmURI,
-			"--openid-provider-timeout=120s",
-			"--listen=" + "0.0.0.0:" + portNum,
-			"--client-id=" + umaTestClient,
-			"--client-secret=" + umaTestClientSecret,
-			"--upstream-url=" + server.URL,
-			"--no-redirects=true",
-			"--enable-uma=true",
-			"--enable-uma-method-scope=true",
-			"--skip-access-token-clientid-check=true",
-			"--skip-access-token-issuer-check=true",
-			"--openid-provider-retry-count=30",
-		}
-
-		fwdProxyArgs := []string{
-			"--discovery-url=" + idpRealmURI,
-			"--openid-provider-timeout=120s",
-			"--listen=" + "0.0.0.0:" + fwdPortNum,
-			"--client-id=" + testClient,
-			"--client-secret=" + testClientSecret,
-			"--enable-uma=true",
-			"--enable-uma-method-scope=true",
-			"--enable-forwarding=true",
-			"--enable-authorization-header=true",
-			"--forwarding-grant-type=client_credentials",
-			"--skip-access-token-clientid-check=true",
-			"--skip-access-token-issuer-check=true",
-			"--openid-provider-retry-count=30",
-		}
-
-		osArgs = append(osArgs, proxyArgs...)
-		startAndWait(portNum, osArgs)
-		fwdOsArgs = append(fwdOsArgs, fwdProxyArgs...)
-		startAndWait(fwdPortNum, fwdOsArgs)
-	})
-
-	When("Accessing resource, where user is allowed to access and then not allowed resource", func() {
-		It("should login with client secret, don't access forbidden resource", func(ctx context.Context) {
-			rClient := resty.New().SetRedirectPolicy(resty.NoRedirectPolicy())
-			rClient.SetProxy(fwdProxyAddress)
-			resp, err := rClient.R().Get(proxyAddress + umaFwdMethodAllowedPath)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusOK))
-
-			body := resp.Body()
-			Expect(strings.Contains(string(body), umaFwdMethodAllowedPath)).To(BeTrue())
-
-			By("Accessing resource without access for client id")
-			resp, err = rClient.R().Get(proxyAddress + umaAllowedPath)
-			body = resp.Body()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusForbidden))
-			Expect(strings.Contains(string(body), umaAllowedPath)).To(BeFalse())
-
-			By("Accessing not allowed method")
-			resp, err = rClient.R().Post(proxyAddress + umaFwdMethodAllowedPath)
-			body = resp.Body()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusForbidden))
-			Expect(strings.Contains(string(body), umaFwdMethodAllowedPath)).To(BeFalse())
-		})
-	})
-})
-
-var _ = Describe("UMA no-redirects authorization with forwarding direct access grant", func() {
-	var portNum string
-	var proxyAddress string
-	var fwdPortNum string
-	var fwdProxyAddress string
-
-	BeforeEach(func() {
-		server := httptest.NewServer(&testsuite.FakeUpstreamService{})
-		portNum = generateRandomPort()
-		fwdPortNum = generateRandomPort()
-		proxyAddress = "http://localhost:" + portNum
-		fwdProxyAddress = "http://localhost:" + fwdPortNum
-		osArgs := []string{os.Args[0]}
-		fwdOsArgs := []string{os.Args[0]}
-		proxyArgs := []string{
-			"--discovery-url=" + idpRealmURI,
-			"--openid-provider-timeout=120s",
-			"--listen=" + "0.0.0.0:" + portNum,
-			"--client-id=" + umaTestClient,
-			"--client-secret=" + umaTestClientSecret,
-			"--upstream-url=" + server.URL,
-			"--no-redirects=true",
-			"--enable-uma=true",
-			"--enable-uma-method-scope=true",
-			"--skip-access-token-clientid-check=true",
-			"--skip-access-token-issuer-check=true",
-			"--openid-provider-retry-count=30",
-		}
-
-		fwdProxyArgs := []string{
-			"--discovery-url=" + idpRealmURI,
-			"--openid-provider-timeout=120s",
-			"--listen=" + "0.0.0.0:" + fwdPortNum,
-			"--client-id=" + testClient,
-			"--client-secret=" + testClientSecret,
-			"--forwarding-username=" + testUser,
-			"--forwarding-password=" + testPass,
-			"--enable-uma=true",
-			"--enable-uma-method-scope=true",
-			"--enable-forwarding=true",
-			"--enable-authorization-header=true",
-			"--skip-access-token-clientid-check=true",
-			"--skip-access-token-issuer-check=true",
-			"--openid-provider-retry-count=30",
-		}
-
-		osArgs = append(osArgs, proxyArgs...)
-		startAndWait(portNum, osArgs)
-		fwdOsArgs = append(fwdOsArgs, fwdProxyArgs...)
-		startAndWait(fwdPortNum, fwdOsArgs)
-	})
-
-	When("Accessing resource, where user is allowed to access and then not allowed resource", func() {
-		It("should login with user/password, don't access forbidden resource", func(ctx context.Context) {
-			rClient := resty.New().SetRedirectPolicy(resty.NoRedirectPolicy())
-			rClient.SetProxy(fwdProxyAddress)
-			resp, err := rClient.R().Get(proxyAddress + umaMethodAllowedPath)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusOK))
-
-			body := resp.Body()
-			Expect(strings.Contains(string(body), umaMethodAllowedPath)).To(BeTrue())
-			Expect(resp.Header().Get(constant.UMAHeader)).NotTo(BeEmpty())
-
-			By("Repeating access to allowed resource, we verify that uma was saved and reused")
-			resp, err = rClient.R().Get(proxyAddress + umaMethodAllowedPath)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusOK))
-
-			body = resp.Body()
-			GinkgoLogr.Info(string(body))
-			Expect(strings.Contains(string(body), umaMethodAllowedPath)).To(BeTrue())
-			Expect(resp.Header().Get(constant.UMAHeader)).To(BeEmpty())
-			// as first request should return uma token in header, it should be
-			// saved in forwarding rpt structure and sent also in this request
-			// so we should see it in response body
-			Expect(strings.Contains(string(body), constant.UMAHeader)).To(BeTrue())
-
-			By("Accessing resource without access for user")
-			resp, err = rClient.R().Get(proxyAddress + umaForbiddenPath)
-			body = resp.Body()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusForbidden))
-			Expect(strings.Contains(string(body), umaForbiddenPath)).To(BeFalse())
-
-			By("Accessing not allowed method")
-			resp, err = rClient.R().Post(proxyAddress + umaMethodAllowedPath)
-			body = resp.Body()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(resp.StatusCode()).To(Equal(http.StatusForbidden))
-			Expect(strings.Contains(string(body), umaMethodAllowedPath)).To(BeFalse())
-		})
-	})
-})
